requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
supabase==2.0.2 